+++ internal/updater/updater.go
@@ -16,7 +16,7 @@ import (
 )
 
 // Version is set at build time via ldflags
-var Version = "6.1.6"
+var Version = "6.1.7"
 
 // GitHub repo info
 const (